Split the tests out a bit.
